The basis to start a pipeline to convert data to an XML file.

This method simply requires the name to save the XML file as and optionally the model interface "header" file that defines the transformation lookup to go from tabular data to hierarchical. The result of this function should be used in a dplyr pipeline with one or more calls to add_xml_data to add the data to convert and finally ending with run_xml_conversion to run the conversion. Please see the wiki for more details https://github.com/JGCRI/gcamdata/wiki/XML-Conversion

create_xml(xml_file, mi_header = NULL)

Arguments

xml_file

The file name to save the XML file to

mi_header

The model interface "header" file name. This will default to the one included in this package

Value

A "data structure" to hold the various parts needed to run the model interface CSV to XML conversion.