module_aglu_L221.land_input_1

Produce L221.LN0_Logit, L221.LN0_Land, L221.LN0_SoilTimeScale, L221.LN1_ValueLogit, L221.LN1_HistUnmgdAllocation, L221.LN1_UnmgdAllocation, and L221.LN1_UnmgdCarbon.

module_aglu_L221.land_input_1(command, ...)

Arguments

command

API command to execute

...

other optional parameters, depending on command

Value

Depends on command: either a vector of required inputs, a vector of output names, or (if command is "MAKE") all the generated outputs: curr_table$data, L221.LN0_Logit, L221.LN0_Land, L221.LN0_SoilTimeScale, curr_table$data, L221.LN1_ValueLogit, L221.LN1_HistUnmgdAllocation, L221.LN1_UnmgdAllocation, L221.LN1_UnmgdCarbon. The corresponding file in the original data system was L221.land_input_1.R (aglu level2).

Details

"L221.LN0_Logit Logit exponent of the top-level (zero) land nest by region. AgLu regions are given externally defined constant logit information."

"L221.LN0_Land Total regional land allocation in top level nest. L125 land cover data is assigned proper GCAM region names and nest id."

"L221.LN0_SoilTimeScale Soil time scale by region. External soil time scale assumptions for each AGLU region."

"L221.LN1_ValueLogit Unmanaged land value by region and GLU, and logit exponent of first nest. L131 land value data is joined with LandNode Logit assumptions.A minimum value is set to ensure all regins have a nonzero land value."

"L221.LN1_HistUnmgdAllocation Historical land cover fora unmanaged land (LT_GLU) in the first nest by region. Historical land cover for unmanaged land in the first nest, from L125 land cover data."

"L221.LN1_UnmgdAllocation Land cover in the model base periods for unmanaged land (LT_GLU) in the first nest by region. Land cover in the model base periods for unmanaged land in the first nest, from L125 land cover data."

"L221.LN1_UnmgdCarbon Carbon content for unmanaged land (LT_GLU) in first nest by region. Carbon content info for unmanaged land in the first nest including soil and vegetative carbon, from L125 land cover data, L121 carbon content data, and GCAMLandLeaf_CdensityLT assumptions."

Author

ACS August 2017