module_aglu_L2231.land_input_3_irr

Produce L2231.LN3_Logit, L2231.LN3_HistUnmgdAllocation, L2231.LN3_UnmgdAllocation, L2231.LN3_HistMgdAllocation_noncrop, L2231.LN3_MgdAllocation_noncrop, L2231.LN3_UnmgdCarbon, L2231.LN3_MgdCarbon_noncrop, L2231.LN3_NoEmissCarbon, L2231.LN3_NodeCarbon, and protected lands related outputs: L2231.LN3_HistUnmgdAllocation_noprot, L2231.LN3_UnmgdAllocation_noprot, L2231.LN1_HistUnmgdAllocation_prot, L2231.LN1_UnmgdAllocation_prot, L2231.LN1_UnmgdCarbon_prot, L2231.LN1_Logit_prot.

module_aglu_L2231.land_input_3_irr(command, ...)

Arguments

command

API command to execute

...

other optional parameters, depending on command

Value

Depends on command: either a vector of required inputs, a vector of output names, or (if command is "MAKE") all the generated outputs: L2231.LN3_Logit, L2231.LN3_HistUnmgdAllocation, L2231.LN3_UnmgdAllocation, L2231.LN3_NoEmissCarbon, L2231.LN3_NodeCarbon, L2231.LN3_HistMgdAllocation_noncrop, L2231.LN3_MgdAllocation_noncrop, L2231.LN3_UnmgdCarbon, L2231.LN3_MgdCarbon_noncrop. The corresponding file in the original data system was L2231.land_input_3_irr.R (aglu level2).

Details

"L2231.LN3_Logit: Logit exponent of the third land nest by region. AgLU regions are given externally defined constant logit information."

"L2231.LN3_HistUnmgdAllocation: Historical land cover for unmanaged land (LT_GLU) in the third nest by region. Historical land cover for unmanaged land in the third nest, from L125 land cover data."

"L2231.LN3_UnmgdAllocation: Land cover in the model base periods for unmanaged land (LT_GLU) in the third nest by region. Land cover in the model base periods for unmanaged land in the third nest, from L125 land cover data."

"L2231.LN3_HistMgdAllocation_noncrop: Historical land cover for non-crop managed land (LT_GLU) in the third nest by region. Historical land cover for non-crop managed land in the third nest, from L125 land cover data."

"L2231.LN3_MgdAllocation_noncrop,: Land cover in the model base periods for non-crop managed land (LT_GLU) in the third nest by region. Land cover in the model base periods for non-crop managed land in the third nest, from L125 land cover data."

"L2231.LN3_UnmgdCarbon: Carbon content for unmanaged land (LT_GLU) in third nest by region. Carbon content info for unmanaged land in the third nest including soil and vegetative carbon, from L125 land cover data, L121 carbon content data, and GCAMLandLeaf_CdensityLT assumptions."

"L2231.LN3_MgdCarbon_noncrop: Carbon content for non-crop managed land (LT_GLU) in third nest by region. Carbon content info for non-crop managed land in the third nest including soil and vegetative carbon, from L125 land cover data, L121 carbon content data, and GCAMLandLeaf_CdensityLT assumptions."

"L2231.LN3_NoEmissCarbon: Sets the no-emiss-carbon-calc as the type of carbon to use in forest leaves by region."

"L2231.LN3_NodeCarbon: Sets the node-carbon-calc to drive the carbon calc between forest leaves, by region, and places the node carbon calc in the node just above the leaves."

"L2231.LN3_HistUnmgdAllocation_noprot: Historical unmanaged land data from L223.LN3_HistUnmgdAllocation is multiplied by specified fraction to give unprotected land allocation in the third nest. OtherArableLands omitted."

"L2231.LN3_UnmgdAllocation_noprot: Unmanaged land data from L223.LN3_UnmgdAllocation is multiplied by specified fraction to give "unprotected land allocation in the third nest. OtherArableLands omitted."

"L2231.LN1_HistUnmgdAllocation_prot: Historical unmanaged land data from L223.LN3_HistUnmgdAllocation is multiplied by specified fraction to give protected land allocation in the first nest. OtherArableLands omitted."

"L2231.LN1_UnmgdAllocation_prot: Unmanaged land data from L223.LN3_UnmgdAllocation is multiplied by specified fraction to give protected land allocation in the first nest. OtherArableLands omitted."

" L2231.LN1_UnmgdCarbon_prot: Carbon content info for protected unmanaged land (LT_GLU) in the first nest including soil and vegetative carbon."

"L2231.LN1_Logit_prot: Logit info for protected lands in the first land nest by region"

Author

ACS September 2017