module_aglu_LB122.LC_R_Cropland_Yh_GLU

Integrate disparate data sources for land cover and harvested area from FAO, Monfreda, and Hyde in order to calculate cropland cover by specific crop type, other arable land, harvested:cropped ratio, economic yield (i.e., production per unit cropland area per year, rather than production per unit area per harvest), and "extra" cropland that needs to be taken from other land use types. All data tables are written out at the GCAM region-GLU level for each historical year, including carbon cycle model spinup years.

module_aglu_LB122.LC_R_Cropland_Yh_GLU(command, ...)

Arguments

command

API command to execute

...

other optional parameters, depending on command

Value

Depends on command: either a vector of required inputs, a vector of output names, or (if command is "MAKE") all the generated outputs: L122.ag_HA_to_CropLand_R_Y_GLU, L122.ag_EcYield_kgm2_R_C_Y_GLU,

L122.LC_bm2_R_OtherArableLand_Yh_GLU, L122.LC_bm2_R_ExtraCropLand_Yh_GLU,

L122.LC_bm2_R_HarvCropLand_C_Yh_GLU, L122.LC_bm2_R_HarvCropLand_Yh_GLU. The corresponding file in the original data system was LB122.LC_R_Cropland_Yh_GLU.R (aglu level1).

Details

First, Hyde cropland cover is aggregated to the region-GLU level in each year, and supplemented with Monfreda harvested area information when necessary. Second, FAO fallowland, cropland, and harvested area are aggregated from the country level to the region level, and used to calculate the "residual" cropland in each land use region: total cropland (Hyde) - known fallow (FAO) - harvested crops (FAO/Monfreda). The harvested area : cropland ratio (HA:CL) is computed as (sum of harvested area) / (cropland - fallow land). Where HA:CL < 1, HA:CL is set to 1, and the balance (the "residual" above) is added to OtherArableLand. Where HA:CL > 3, HA:CL is set to 3, and additional cropland (ExtraCropLand) is written out, to be taken from other land use types (determined in different code chunks).

Author

ACS April 2017