module_emissions_L123.bcoc_awb_R_S_T_Y

Produces BC and OC AWB emission factors for each Ag production technology.

module_emissions_L123.bcoc_awb_R_S_T_Y(command, ...)

Arguments

command

API command to execute

...

other optional parameters, depending on command

Value

Depends on command: either a vector of required inputs, a vector of output names, or (if command is "MAKE") all the generated outputs: L123.bcoc_tgmt_R_awb_2000. The corresponding file in the original data system was L123.bcoc_awb_R_S_T_Y.R (emissions level1).

Details

Use RCP year 2000 BC and OC emissions from agricultural waste burning on fields (AWB) to calculate emission factors for each ag production technology. Allocates regional AWB emissions to GCAM technology according to above-ground non-harvested biomass.

Author

SJS May 2017