module_energy_L244.building_det

Creates level2 data for the building sector.

module_energy_L244.building_det(command, ...)

Arguments

command

API command to execute

...

other optional parameters, depending on command

Value

Depends on command: either a vector of required inputs, a vector of output names, or (if command is "MAKE") all the generated outputs: L244.SubregionalShares, L244.PriceExp_IntGains, L244.Floorspace, L244.DemandFunction_serv, L244.DemandFunction_flsp, L244.Satiation_flsp, L244.SatiationAdder, L244.ThermalBaseService, L244.GenericBaseService, L244.ThermalServiceSatiation, L244.GenericServiceSatiation, L244.Intgains_scalar, L244.ShellConductance_bld, L244.Supplysector_bld, L244.FinalEnergyKeyword_bld, L244.SubsectorShrwt_bld, L244.SubsectorShrwtFllt_bld, L244.SubsectorInterp_bld, L244.SubsectorInterpTo_bld, L244.SubsectorLogit_bld, L244.FuelPrefElast_bld, L244.StubTech_bld, L244.StubTechEff_bld, L244.StubTechCalInput_bld, L244.StubTechIntGainOutputRatio, L244.GlobalTechShrwt_bld, L244.GlobalTechCost_bld, L244.DeleteGenericService, L244.Satiation_flsp_SSP1, L244.SatiationAdder_SSP1, L244.GenericServiceSatiation_SSP1, L244.Satiation_flsp_SSP2, L244.SatiationAdder_SSP2, L244.GenericServiceSatiation_SSP2, L244.Satiation_flsp_SSP3, L244.SatiationAdder_SSP3, L244.GenericServiceSatiation_SSP3, L244.FuelPrefElast_bld_SSP3, L244.Satiation_flsp_SSP4, L244.SatiationAdder_SSP4, L244.GenericServiceSatiation_SSP4, L244.FuelPrefElast_bld_SSP4, L244.Satiation_flsp_SSP5, L244.SatiationAdder_SSP5, L244.GenericServiceSatiation_SSP5, L244.FuelPrefElast_bld_SSP15, L244.DeleteThermalService, L244.HDDCDD_A2_CCSM3x, L244.HDDCDD_A2_HadCM3, L244.HDDCDD_B1_CCSM3x, L244.HDDCDD_B1_HadCM3, L244.HDDCDD_constdd_no_GCM and L244.GompFnParam. The corresponding file in the original data system was L244.building_det.R (energy level2).

Details

Creates level2 data for the building sector.

Author

RLH September 2017